feat: 수업 마감&삭제 Class -> MyClass
